Refactor Rates Trends
