+++ src/components/rates/history/RatesTrends.tsx
@@ -1,24 +1,6 @@
 import { useStore } from "@nanostores/react";
 import { TrendingUp } from "lucide-react";
 import { useEffect, useMemo } from "react";
-import { LenderLogo } from "@/components/lenders/LenderLogo";
-import { LenderSelector } from "@/components/lenders/LenderSelector";
-import { Label } from "@/components/ui/label";
-import {
-	Select,
-	SelectContent,
-	SelectItem,
-	SelectTrigger,
-	SelectValue,
-} from "@/components/ui/select";
-import {
-	Table,
-	TableBody,
-	TableCell,
-	TableHead,
-	TableHeader,
-	TableRow,
-} from "@/components/ui/table";
 import type { Lender } from "@/lib/schemas/lender";
 import type {
 	RatesHistoryFile,
@@ -28,46 +10,16 @@ import { getRateTimeSeries } from "@/lib/stores/rates/rates-history";
 import {
 	$trendsFilter,
 	$trendsSelectedLenders,
-	setTrendsFilter,
 	setTrendsSelectedLenders,
 } from "@/lib/stores/rates/rates-history-filters";
-import { formatShortMonthYearFromString } from "@/lib/utils/date";
 import { RatesTrendChart } from "./RatesTrendChart";
+import { TrendsFilters } from "./TrendsFilters";
+import { type RateStat, TrendsStatsTable } from "./TrendsStatsTable";
 
-interface TrendChartsProps {
+interface RatesTrendsProps {
 	historyData: Map<string, RatesHistoryFile>;
 	lenders: Lender[];
 }
-
-// Common rate types to filter by
-const RATE_TYPES = [
-	{ value: "all", label: "All Rate Types" },
-	{ value: "fixed-1", label: "1-Year Fixed" },
-	{ value: "fixed-2", label: "2-Year Fixed" },
-	{ value: "fixed-3", label: "3-Year Fixed" },
-	{ value: "fixed-4", label: "4-Year Fixed" },
-	{ value: "fixed-5", label: "5-Year Fixed" },
-	{ value: "fixed-7", label: "7-Year Fixed" },
-	{ value: "fixed-10", label: "10-Year Fixed" },
-	{ value: "variable", label: "Variable" },
-];
-
-// LTV ranges to filter by
-const LTV_RANGES = [
-	{ value: "all", label: "All LTV" },
-	{ value: "50", label: "Up to 50% LTV" },
-	{ value: "60", label: "Up to 60% LTV" },
-	{ value: "70", label: "Up to 70% LTV" },
-	{ value: "80", label: "Up to 80% LTV" },
-	{ value: "90", label: "Up to 90% LTV" },
-];
-
-// Buyer categories (Primary Residence and BTL are mutually exclusive mortgage types)
-const BUYER_CATEGORIES = [
-	{ value: "pdh", label: "Primary Residence" },
-	{ value: "btl", label: "Buy to Let" },
-	{ value: "all", label: "All Buyers" },
-] as const;
 
 // Buyer types that belong to each category
 const PDH_BUYER_TYPES = ["ftb", "mover", "switcher-pdh"] as const;
@@ -82,7 +34,7 @@ function getRateTypeKey(rate: { type: string; fixedTerm?: number }): string {
 	return "other";
 }
 
-export function RatesTrendCharts({ historyData, lenders }: TrendChartsProps) {
+export function RatesTrends({ historyData, lenders }: RatesTrendsProps) {
 	const filter = useStore($trendsFilter);
 	const selectedLenders = useStore($trendsSelectedLenders);
 
@@ -254,7 +206,7 @@ export function RatesTrendCharts({ historyData, lenders }: TrendChartsProps) {
 	}, [timeSeries]);
 
 	// Calculate min/max/current stats for each rate
-	const rateStats = useMemo(() => {
+	const rateStats = useMemo((): RateStat[] => {
 		return timeSeries.map((series) => {
 			const lender = lenders.find((l) => l.id === series.lenderId);
 			const points = series.dataPoints;
@@ -293,96 +245,10 @@ export function RatesTrendCharts({ historyData, lenders }: TrendChartsProps) {
 		});
 	}, [timeSeries, lenders]);
 
-	// Parse LTV filter value
-	const ltvValue = filter.ltvRange ? String(filter.ltvRange[1]) : "all";
-
 	return (
 		<div className="space-y-4">
 			{/* Filters */}
-			<div className="flex flex-wrap items-start gap-4 p-4 rounded-lg bg-muted/50">
-				{/* Rate Type Filter */}
-				<div className="space-y-1.5">
-					<Label className="text-xs">Rate Type</Label>
-					<Select
-						value={filter.rateType ?? "all"}
-						onValueChange={(v) =>
-							setTrendsFilter({ rateType: v === "all" ? null : v })
-						}
-					>
-						<SelectTrigger className="w-[140px]">
-							<SelectValue />
-						</SelectTrigger>
-						<SelectContent>
-							{RATE_TYPES.map((type) => (
-								<SelectItem key={type.value} value={type.value}>
-									{type.label}
-								</SelectItem>
-							))}
-						</SelectContent>
-					</Select>
-				</div>
-
-				{/* LTV Filter */}
-				<div className="space-y-1.5">
-					<Label className="text-xs">Max LTV</Label>
-					<Select
-						value={ltvValue}
-						onValueChange={(v) =>
-							setTrendsFilter({
-								ltvRange: v === "all" ? null : [0, Number.parseInt(v, 10)],
-							})
-						}
-					>
-						<SelectTrigger className="w-[120px]">
-							<SelectValue />
-						</SelectTrigger>
-						<SelectContent>
-							{LTV_RANGES.map((ltv) => (
-								<SelectItem key={ltv.value} value={ltv.value}>
-									{ltv.label}
-								</SelectItem>
-							))}
-						</SelectContent>
-					</Select>
-				</div>
-
-				{/* Buyer Category Filter */}
-				<div className="space-y-1.5">
-					<Label className="text-xs">Buyer Type</Label>
-					<Select
-						value={filter.buyerCategory}
-						onValueChange={(v) =>
-							setTrendsFilter({
-								buyerCategory: v as "all" | "pdh" | "btl",
-							})
-						}
-					>
-						<SelectTrigger className="w-[180px]">
-							<SelectValue />
-						</SelectTrigger>
-						<SelectContent>
-							{BUYER_CATEGORIES.map((cat) => (
-								<SelectItem key={cat.value} value={cat.value}>
-									{cat.label}
-								</SelectItem>
-							))}
-						</SelectContent>
-					</Select>
-				</div>
-
-				{/* Lender Selection */}
-				<div className="space-y-1.5">
-					<Label className="text-xs">Lenders</Label>
-					<LenderSelector
-						lenders={lenders}
-						value={selectedLenders}
-						onChange={setTrendsSelectedLenders}
-						multiple
-						placeholder="Select lenders"
-						className="w-[260px]"
-					/>
-				</div>
-			</div>
+			<TrendsFilters lenders={lenders} />
 
 			{/* Chart */}
 			{timeSeries.length === 0 ? (
@@ -409,79 +275,7 @@ export function RatesTrendCharts({ historyData, lenders }: TrendChartsProps) {
 					/>
 
 					{/* Rate Stats Table */}
-					{rateStats.length > 0 && (
-						<div className="rounded-lg border overflow-hidden">
-							<Table>
-								<TableHeader>
-									<TableRow className="bg-muted/50">
-										<TableHead className="font-medium">Lender</TableHead>
-										<TableHead className="font-medium">Rate</TableHead>
-										<TableHead className="text-right font-medium">
-											Min
-										</TableHead>
-										<TableHead className="text-right font-medium">
-											Max
-										</TableHead>
-										<TableHead className="text-right font-medium">
-											Current
-										</TableHead>
-									</TableRow>
-								</TableHeader>
-								<TableBody>
-									{rateStats.map((stat) => (
-										<TableRow key={stat.rateId}>
-											<TableCell>
-												<div className="flex items-center gap-2">
-													<LenderLogo lenderId={stat.lenderId} size={36} />
-													<span className="font-medium">{stat.lenderName}</span>
-												</div>
-											</TableCell>
-											<TableCell className="text-muted-foreground">
-												{stat.rateName}
-											</TableCell>
-											<TableCell className="text-right">
-												{stat.min ? (
-													<div>
-														<div className="font-mono">
-															{stat.min.rate.toFixed(2)}%
-														</div>
-														<div className="text-xs text-muted-foreground">
-															{formatShortMonthYearFromString(stat.min.date)}
-														</div>
-													</div>
-												) : (
-													"—"
-												)}
-											</TableCell>
-											<TableCell className="text-right">
-												{stat.max ? (
-													<div>
-														<div className="font-mono">
-															{stat.max.rate.toFixed(2)}%
-														</div>
-														<div className="text-xs text-muted-foreground">
-															{formatShortMonthYearFromString(stat.max.date)}
-														</div>
-													</div>
-												) : (
-													"—"
-												)}
-											</TableCell>
-											<TableCell className="text-right">
-												{stat.current ? (
-													<div className="font-mono font-medium">
-														{stat.current.rate.toFixed(2)}%
-													</div>
-												) : (
-													"—"
-												)}
-											</TableCell>
-										</TableRow>
-									))}
-								</TableBody>
-							</Table>
-						</div>
-					)}
+					<TrendsStatsTable rateStats={rateStats} />
 				</div>
 			)}
 		</div>
